requirements.txt
# required python modules to run task.py
lxml
pillow